requirements.txt
absl-py==0.13.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
colorama==0.4.4
cycler==0.10.0
EasyProcess==0.3
entrypoint2==0.2.4
fire==0.4.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.7.0
joblib==1.0.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.3
mss==6.1.0
numpy==1.19.5
oauthlib==3.1.1
onnx==1.5.0
onnxconverter-common==1.8.1
onnxmltools==1.5.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyscreenshot==3.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
simplejson==3.17.5
sip==4.19.8
six==1.15.0
skl2onnx==1.9.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tf2onnx==1.5.3
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
tqdm==4.62.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
winmltools==1.5.1
wrapt==1.12.1
zipp==3.5.0
